requirements.txt
# GitHub API
PyGithub>=2.1.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.2

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0

# Configuration
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.65.0
